Basics of Notes
This branch adds a 'add a note' and 'notes' page. The add a note page allows you to add a note that is stored in the database and is linked to the user you are signed in as. You can not view either of these pages if you are not signed in. The functionallity to see notes you have made is not as of yet implemented on the your notes page.